requirements.txt
# Telegram Bot
python-telegram-bot==21.5
python-telegram-bot[job-queue]==21.5

# Remnawave API SDK (Official)
remnawave==2.1.17

# Configuration
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# Logging
loguru==0.7.2

# Utils
python-dateutil==2.9.0.post0
httpx==0.27.2

# Caching (optional)
redis==5.1.1
hiredis==3.0.0

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.8.0
flake8==7.1.1
mypy==1.11.2
